Refactor routes and controllers, reintroduce rate limiting
