+++ src/db_utils.py
@@ -1,6 +1,5 @@
-def init_db(app):
-    from models.models import db
+from models.models import db
 
-    db.init_app(app)
+def init_db(app):
     with app.app_context():
         db.create_all()
