+++ src/templates/layout.html
@@ -23,9 +23,9 @@
       <nav>
         <ul class="tabs">
           <li><a href="{{ url_for('index') }}" class="tab-link" data-tab="home">Home</a></li>
-          <li><a href="{{ url_for('visualize') }}" class="tab-link" data-tab="visualize">Visualization</a></li>
-          <li><a href="{{ url_for('previous_projects') }}" class="tab-link" data-tab="previous-projects">Previous Projects</a></li>
-          <li><a href="{{ url_for('update_project') }}" class="tab-link" data-tab="update-project">Update Project</a></li>
+          <li><a href="{{ url_for('visualize_route') }}" class="tab-link" data-tab="visualize">Visualization</a></li>
+          <li><a href="{{ url_for('previous_projects_route') }}" class="tab-link" data-tab="previous-projects">Previous Projects</a></li>
+          <li><a href="{{ url_for('update_project_route') }}" class="tab-link" data-tab="update-project">Update Project</a></li>
         </ul>
       </nav>
     </header>
